service layer clean & dto folder moved
