+++ packages/headless/src/modules/wishlist/rest/dto/deleteAllWishlistItems.dto.ts
