+++ packages/headless/src/modules/wishlist/rest/dto/deleteWishlist.dto.ts
