+++ packages/headless/src/modules/wishlist/rest/dto/deleteWishlistItem.dto.ts
