+++ packages/headless/src/modules/wishlist/rest/dto/getUserWishlist.dto.ts
