+++ packages/headless/src/modules/wishlist/rest/dto/updateWishlistItem.dto.ts
